Recording/Main: Removed xml/Settings dependency from VirtualBox client side API (VBoxC) [macOS build fix]. bugref:10943
svn:sync-xref-src-repo-rev: r169810

+++ src/VBox/Main/src-all/RecordingUtils.cpp
@@ -1,4 +1,4 @@
-/* $Id: RecordingUtils.cpp 110354 2025-07-22 16:25:44Z andreas.loeffler@oracle.com $ */
+/* $Id: RecordingUtils.cpp 110355 2025-07-22 16:43:55Z andreas.loeffler@oracle.com $ */
 /** @file
  * Recording utility code.
  */
@@ -84,8 +84,7 @@ inline bool recordingUtilsColorConvWriteYUV420p(uint8_t *aDstBuf, unsigned aDstW
                 uint8_t g = aSrcBuf[4 * i + 1];
                 uint8_t r = aSrcBuf[4 * i + 2];
 
-                aDstBuf[i++] = (uint8_t)CALC_Y(r, g, b);
-
+                aDstBuf[i++]    =          CALC_Y(r, g, b);
                 aDstBuf[upos++] = (uint8_t)CALC_U(r, g, b);
                 aDstBuf[vpos++] = (uint8_t)CALC_V(r, g, b);
 
@@ -93,7 +92,7 @@ inline bool recordingUtilsColorConvWriteYUV420p(uint8_t *aDstBuf, unsigned aDstW
                 g = aSrcBuf[4 * i + 1];
                 r = aSrcBuf[4 * i + 2];
 
-                aDstBuf[i++] = (uint8_t)CALC_Y(r, g, b);
+                aDstBuf[i++] = CALC_Y(r, g, b);
             }
         }
         else
@@ -183,8 +182,7 @@ void RecordingUtilsConvBGRA32ToYUVI420(uint8_t *paDst, uint32_t uDstWidth, uint3
                 uint8_t g = paSrc[4 * i + 1];
                 uint8_t r = paSrc[4 * i + 2];
 
-                paDst[i++] = (uint8_t)CALC_Y(r, g, b);
-
+                paDst[i++]    =          CALC_Y(r, g, b);
                 paDst[upos++] = (uint8_t)CALC_U(r, g, b);
                 paDst[vpos++] = (uint8_t)CALC_V(r, g, b);
 
@@ -192,7 +190,7 @@ void RecordingUtilsConvBGRA32ToYUVI420(uint8_t *paDst, uint32_t uDstWidth, uint3
                 g = paSrc[4 * i + 1];
                 r = paSrc[4 * i + 2];
 
-                paDst[i++] = (uint8_t)CALC_Y(r, g, b);
+                paDst[i++] = CALC_Y(r, g, b);
             }
         }
         else
@@ -203,7 +201,7 @@ void RecordingUtilsConvBGRA32ToYUVI420(uint8_t *paDst, uint32_t uDstWidth, uint3
                 uint8_t const g = paSrc[4 * i + 1];
                 uint8_t const r = paSrc[4 * i + 2];
 
-                paDst[i++] = (uint8_t)CALC_Y(r, g, b);
+                paDst[i++] = CALC_Y(r, g, b);
             }
         }
     }
@@ -267,7 +265,7 @@ void RecordingUtilsConvBGRA32ToYUVI420Ex(uint8_t *paDst, uint32_t uDstX, uint32_
             size_t const offY  = uDstY * uDstWidth + uDstXCur;
             size_t const offUV = (uDstY / 2) * (uDstWidth / 2) + (uDstXCur / 2) + uDstWidth * uDstHeight;
 
-            paDst[offY]                               = (uint8_t)CALC_Y(r, g, b);
+            paDst[offY]                               =          CALC_Y(r, g, b);
             paDst[offUV]                              = (uint8_t)CALC_U(r, g, b);
             paDst[offUV + uDstWidth * uDstHeight / 4] = (uint8_t)CALC_V(r, g, b);
